+++ src/app/pages/list/list.spec.ts
@@ -17,7 +17,7 @@ describe('ListPage', () => {
   it('should render without error', () => {
     fixture.detectChanges();
     const el = fixture.nativeElement;
-    expect(el.querySelector('.list-page')).toBeTruthy();
-    expect(el.querySelector('.list-title').textContent).toContain('Projects');
+    expect(el.querySelector('.full-width-layout')).toBeTruthy();
+    expect(el.querySelector('.full-width-layout-title').textContent).toContain('Projects');
   });
 });